+++ CallBackJob.go
@@ -10,7 +10,7 @@ var ERR_CALLBACK_PANIC = errors.New("Callback Panic")
 
 type CallBackJob struct {
 	timeout int64
-	events  int16
+	events  uint32
 	onEvent func(event *CallbackEvent)
 	worker  *Worker
 	fd      int32
@@ -19,7 +19,7 @@ type CallBackJob struct {
 	jobId   int64
 }
 
-func NewJobFromFdT(fd int32, watchEvents int16, timeout int64, cb func(*CallbackEvent)) (job *CallBackJob) {
+func NewJobFromFdT(fd int32, watchEvents uint32, timeout int64, cb func(*CallbackEvent)) (job *CallBackJob) {
 	job = &CallBackJob{
 		onEvent: cb,
 		timeout: timeout,
@@ -31,7 +31,7 @@ func NewJobFromFdT(fd int32, watchEvents int16, timeout int64, cb func(*Callback
 	return
 }
 
-func NewJobFromOsFileT(f *os.File, watchEvents int16, timeout int64, cb func(*CallbackEvent)) *CallBackJob {
+func NewJobFromOsFileT(f *os.File, watchEvents uint32, timeout int64, cb func(*CallbackEvent)) *CallBackJob {
 	return NewJobFromFdT(int32(f.Fd()), watchEvents, timeout, cb)
 }
 
@@ -50,7 +50,7 @@ func (s *CallBackJob) SetTimeout(t int64) {
 }
 
 // Processes the last epoll events and returns the next flags to use.
-func (s *CallBackJob) ProcessEvents(currentEvents int16, now int64) (watchEevents int16, futureTimeOut int64, EventError error) {
+func (s *CallBackJob) ProcessEvents(currentEvents uint32, now int64) (watchEevents uint32, futureTimeOut int64, EventError error) {
 
 	s.Lock.RLock()
 	defer s.Lock.RUnlock()
@@ -91,7 +91,7 @@ func (s *CallBackJob) safeEvent(config *CallbackEvent) {
 
 // Called to validate the "lastTimeout", should return a futureTimeOut or 0 if there is no timeout.
 // If the Job has timed out TimeOutError should be set to os.ErrDeadlineExceeded.
-func (s *CallBackJob) CheckTimeOut(now int64, lastTimeout int64) (NewEvents int16, futureTimeOut int64, TimeOutError error) {
+func (s *CallBackJob) CheckTimeOut(now int64, lastTimeout int64) (NewEvents uint32, futureTimeOut int64, TimeOutError error) {
 	s.Lock.RLock()
 	defer s.Lock.RUnlock()
 	s.ran = false
@@ -131,7 +131,7 @@ func (s *CallBackJob) CheckTimeOut(now int64, lastTimeout int64) (NewEvents int1
 }
 
 // Sets the current Worker. This method is called when a Job is added to a Worker in the pool.
-func (s *CallBackJob) SetPool(worker *Worker, now int64) (watchEevents int16, futureTimeOut int64, fd int32) {
+func (s *CallBackJob) SetPool(worker *Worker, now int64) (watchEevents uint32, futureTimeOut int64, fd int32) {
 	s.Lock.Lock()
 	defer s.Lock.Unlock()
 	s.ran = false
@@ -164,7 +164,7 @@ func (s *CallBackJob) onRecover(config *CallbackEvent) {
 	}
 }
 
-func (s *CallBackJob) SetEvents(events int16) error {
+func (s *CallBackJob) SetEvents(events uint32) error {
 	s.Lock.Lock()
 	defer s.Lock.Unlock()
 	s.events = events
@@ -185,7 +185,7 @@ func (s *CallBackJob) Release() error {
 	return nil
 }
 
-func (s *CallBackJob) GetSettings() (fd int32, events int16, timeout int64, cb func(*CallbackEvent)) {
+func (s *CallBackJob) GetSettings() (fd int32, events uint32, timeout int64, cb func(*CallbackEvent)) {
 	s.Lock.RLock()
 	defer s.Lock.RUnlock()
 	fd = s.fd
